Refactor ResNet model to support LoRA configurations and add training utilities
- Updated `conv3x3` and `conv1x1` functions to accept LoRA configuration parameters.
- Modified `Bottleneck` and `ResNet` classes to handle LoRA configurations.
- Renamed `load_weight` to `load_weight_lora` for clarity.
- Added new training script `train.py` for various training strategies including full fine-tuning, MetaLR, Conv-Adapters, and LoRA.
- Introduced `ExcelTrainingLogger` for logging training metrics and parameters to Excel.
- Created dataset loaders for Kaggle Brain MRI, ISIC, and PathMNIST datasets.
- Implemented utility functions for training and evaluation in `resnet_utils.py`.
- Removed unused `misc.py` file.
- Deleted outdated model weights and sample images.
- Added temporary training results file for logging experiment outcomes.
