+++ models/conv_adapter_module.py
@@ -4,6 +4,7 @@
 # Defines ConvAdapter block for ResNet from the paper Conv-Adapter: Exploring Parameter Efficient Transfer Learning for ConvNets
 # -------------------------
 
+from models.resnet_base import Bottleneck, resnet50_base
 import torch.nn as nn
 
 class ConvAdapter(nn.Module):
@@ -25,3 +26,27 @@ class ConvAdapter(nn.Module):
     def forward(self, x):
         return x + self.adapter(x)
 
+def add_conv_to_resnet(model, reduction=4):
+    """
+    Adds ConvAdapters to each Bottleneck block of a ResNet model in-place.
+    """
+    for module in model.modules():
+        if isinstance(module, Bottleneck):
+            module.adapter = ConvAdapter(in_channels=module.adapter_in_channels, reduction=reduction)
+    return model
+
+def freeze_conv_encoder(model):
+    """
+    Freezes all ResNet weights except for adapters and the final classification head.
+    """
+    for name, param in model.named_parameters():
+        if 'adapter' not in name and 'fc' not in name:
+            param.requires_grad = False
+    return model
+
+def initialize_conv_model(num_classes, device, reduction):
+    model = resnet50_base(pretrained=True, num_classes=num_classes)
+    add_conv_to_resnet(model, reduction=reduction)
+    freeze_conv_encoder(model)
+    model.to(device)
+    return model
