+++ results/temp_training_results.txt
@@ -1,0 +1,201 @@
+=== Training with Full Fine-Tuning on half of kaggle dataset ===
+Train Loss: 1.6984, Accuracy: 0.4266
+FFT Epoch 1/5 - Val Acc: 0.6410
+Train Loss: 0.9434, Accuracy: 0.6289
+FFT Epoch 2/5 - Val Acc: 0.4786
+Train Loss: 0.9330, Accuracy: 0.6275
+FFT Epoch 3/5 - Val Acc: 0.6211
+Train Loss: 0.8613, Accuracy: 0.6660
+FFT Epoch 4/5 - Val Acc: 0.5897
+Train Loss: 0.7748, Accuracy: 0.7080
+FFT Epoch 5/5 - Val Acc: 0.6382
+FFT Test Acc: 0.6492
+
+=== Training with MetaLR on half of kaggle dataset ===
+MetaLR Epoch 1/5 - Val Acc: 0.2934
+MetaLR Epoch 2/5 - Val Acc: 0.5897
+MetaLR Epoch 3/5 - Val Acc: 0.8319
+MetaLR Epoch 4/5 - Val Acc: 0.6809
+MetaLR Epoch 5/5 - Val Acc: 0.6895
+MetaLR Test Acc: 0.6953
+
+With LORA-C model_________________________________________________
+
+--- Kaggle Brain MRI Dataset Summary ---
+Number of classes: 4
+Train samples: 4494
+Val samples:   1123
+Test samples:  1406
+Class distribution (train):
+[1044, 1036, 1273, 1141]
+Total parameters:     23,516,228
+Trainable parameters: 26,564
+Frozen parameters:    23,489,664
+Trainable ratio:      0.11%
+Frozen ratio:         99.89%
+
+--- Parameter breakdown by module ---
+conv1                | total: 9,408 | trainable: 0 | frozen: 9,408
+bn1                  | total: 128 | trainable: 64 | frozen: 64
+relu                 | total: 0 | trainable: 0 | frozen: 0
+maxpool              | total: 0 | trainable: 0 | frozen: 0
+layer1               | total: 215,808 | trainable: 1,408 | frozen: 214,400
+layer2               | total: 1,219,584 | trainable: 3,584 | frozen: 1,216,000
+layer3               | total: 7,098,368 | trainable: 10,240 | frozen: 7,088,128
+layer4               | total: 14,964,736 | trainable: 11,264 | frozen: 14,953,472
+avgpool              | total: 0 | trainable: 0 | frozen: 0
+fc                   | total: 8,196 | trainable: 4 | frozen: 8,192
+
+Epoch 1/5
+Train Loss: 1.2974, Accuracy: 0.3632
+Validation Accuracy: 0.4550
+
+Epoch 2/5
+Train Loss: 1.1525, Accuracy: 0.4691
+Validation Accuracy: 0.5245
+
+Epoch 3/5
+Train Loss: 1.0415, Accuracy: 0.5592
+Validation Accuracy: 0.6046
+
+Epoch 4/5
+Train Loss: 0.9432, Accuracy: 0.6455
+Validation Accuracy: 0.7106
+
+Epoch 5/5
+Train Loss: 0.8119, Accuracy: 0.7194
+Validation Accuracy: 0.7302
+Test Accuracy: 0.7589
+Model saved to .\resnet50_lora_bs32_lr0.001_ep5.pth
+
+With CONV-ADAPTER___________________________________________________________________
+
+--- Kaggle Brain MRI Dataset Summary ---
+Number of classes: 4
+Train samples: 4494
+Val samples:   1123
+Test samples:  1406
+Class distribution (train):
+[1043, 1061, 1274, 1116]
+
+
+Total parameters:     23,559,949
+Trainable parameters: 51,917
+Frozen parameters:    23,508,032
+Trainable ratio:      0.22%
+Frozen ratio:         99.78%
+
+--- Parameter breakdown by module ---
+conv1                | total: 9,408 | trainable: 0 | frozen: 9,408
+bn1                  | total: 128 | trainable: 0 | frozen: 128
+relu                 | total: 0 | trainable: 0 | frozen: 0
+maxpool              | total: 0 | trainable: 0 | frozen: 0
+layer1               | total: 216,417 | trainable: 609 | frozen: 215,808
+layer2               | total: 1,222,232 | trainable: 2,648 | frozen: 1,219,584
+layer3               | total: 7,112,456 | trainable: 14,088 | frozen: 7,098,368
+layer4               | total: 14,991,112 | trainable: 26,376 | frozen: 14,964,736
+avgpool              | total: 0 | trainable: 0 | frozen: 0
+fc                   | total: 8,196 | trainable: 8,196 | frozen: 0
+
+Epoch 1/5
+Train Loss: 0.5475, Accuracy: 0.7928
+Validation Accuracy: 0.8540
+
+Epoch 2/5
+Train Loss: 0.2494, Accuracy: 0.9090
+Validation Accuracy: 0.9172
+
+Epoch 3/5
+Train Loss: 0.1836, Accuracy: 0.9328
+Validation Accuracy: 0.9225
+
+Epoch 4/5
+Train Loss: 0.1504, Accuracy: 0.9448
+
+Epoch 5/5
+Train Loss: 0.1035, Accuracy: 0.9624
+Validation Accuracy: 0.9537
+
+
+Test Accuracy: 0.9630
+Model saved to .\resnet50_adapters_bs32_lr0.001_red64_ep5.pth
+_______________________________________________________________________________________
+
+Total parameters:     24,162,664
+Trainable parameters: 654,632
+Frozen parameters:    23,508,032
+
+--- Parameter breakdown by module ---
+conv1                | total: 9,408 | trainable: 0 | frozen: 9,408
+bn1                  | total: 128 | trainable: 0 | frozen: 128
+relu                 | total: 0 | trainable: 0 | frozen: 0
+maxpool              | total: 0 | trainable: 0 | frozen: 0
+layer1               | total: 222,852 | trainable: 7,044 | frozen: 215,808
+layer2               | total: 1,254,752 | trainable: 35,168 | frozen: 1,219,584
+layer3               | total: 7,302,176 | trainable: 203,808 | frozen: 7,098,368
+layer4               | total: 15,365,152 | trainable: 400,416 | frozen: 14,964,736
+avgpool              | total: 0 | trainable: 0 | frozen: 0
+fc                   | total: 8,196 | trainable: 8,196 | frozen: 0
+
+Epoch 1/5
+Train Loss: 0.4429, Accuracy: 0.8235
+
+Epoch 2/5
+Train Loss: 0.1838, Accuracy: 0.9335
+
+Epoch 3/5
+Train Loss: 0.0982, Accuracy: 0.9637
+
+Epoch 4/5
+Train Loss: 0.0810, Accuracy: 0.9733
+
+Epoch 5/5
+Train Loss: 0.0342, Accuracy: 0.9882
+
+--------------------------------------------------------------------------------------------------------
+Total parameters:     33,632,644
+Trainable parameters: 10,124,612
+Frozen parameters:    23,508,032
+
+--- Parameter breakdown by module ---
+conv1                | total: 9,408 | trainable: 0 
+bn1                  | total: 128 | trainable: 0 
+relu                 | total: 0 | trainable: 0 
+maxpool              | total: 0 | trainable: 0
+layer1               | total: 316,992 | trainable: 101,184
+layer2               | total: 1,751,552 | trainable: 531,968 
+layer3               | total: 10,267,136 | trainable: 3,168,768 
+layer4               | total: 21,279,232 | trainable: 6,314,496 
+avgpool              | total: 0 | trainable: 0 
+fc                   | total: 8,196 | trainable: 8,196 
+
+Epoch 1/10
+Train Loss: 0.4101, Accuracy: 0.8569
+
+Epoch 2/10
+Train Loss: 0.1592, Accuracy: 0.9497
+
+Epoch 3/10
+Train Loss: 0.0863, Accuracy: 0.9711
+
+Epoch 4/10
+Train Loss: 0.0873, Accuracy: 0.9695
+
+Epoch 5/10
+Train Loss: 0.0493, Accuracy: 0.9824
+
+Epoch 6/10
+Train Loss: 0.0473, Accuracy: 0.9838
+
+Epoch 7/10
+Train Loss: 0.0331, Accuracy: 0.9900
+
+Epoch 8/10
+Train Loss: 0.0250, Accuracy: 0.9924
+
+Epoch 9/10
+Train Loss: 0.0289, Accuracy: 0.9900
+
+Epoch 10/10
+Train Loss: 0.0451, Accuracy: 0.9884
+
